move remaining properties
